+++ webpack.config.js
@@ -1,6 +1,4 @@
-const { DefinePlugin } = require('webpack');
-const { join } = require('path');
-const Visualizer = require('webpack-visualizer-plugin');
+const { StatsWriterPlugin } = require('webpack-stats-plugin');
 
 module.exports = {
   entry: {
@@ -8,18 +6,36 @@ module.exports = {
   },
   externals: ['net'],
   mode: 'production',
+  module: {
+    rules: [
+      {
+        // To speed up bundling, we are limiting Babel to a number of packages which does not publish ES5 bits.
+        test: /\/node_modules\/(botframework-streaming|buffer)\//iu,
+        use: {
+          loader: 'babel-loader',
+          options: {
+            presets: [
+              [
+                '@babel/preset-env',
+                {
+                  modules: 'commonjs'
+                }
+              ]
+            ]
+          }
+        }
+      }
+    ]
+  },
   output: {
-    filename: '[name].js',
     library: 'DirectLine',
-    libraryTarget: 'umd',
-    path: join(__dirname, 'dist')
+    libraryTarget: 'umd'
   },
   plugins: [
-    new DefinePlugin({
-      'process.env': {
-        'VERSION': JSON.stringify(process.env.npm_package_version)
-      }
-    }),
-    new Visualizer()
-  ]
+    new StatsWriterPlugin({
+      filename: 'stats.json',
+      transform: (_, opts) => JSON.stringify(opts.compiler.getStats().toJson({ chunkModules: true }), null, 2)
+    })
+  ],
+  target: ['web', 'es5']
 };